ext: mcux: update struct fields in MacInterface.h
Update mlmeStartReq_tag and mcpsDataReq_tag structures so they
can be directly mapped to buffers.

Remove channelPage field from mlmeStartReq_tag as it is not used
in the MAC start request.

Reorder source and destination fields from mcpsDataReq_tag for
them to be in the same order as in the MAC data request.

Signed-off-by: Bogdan Davidoaia <bogdan.davidoaia@linaro.org>