soc: riscv32: Generate event symbol for both rv32m1 cores
The soc irq handler uses a symbol generated in offsets.c to get the
address of the EVENTx_INTPTPENDCLEAR register. Rework the generation of
this symbol such that the ri5cy gets the register from the EVENT0
instance and the zero riscy core gets the register from the EVENT1
instance.

Signed-off-by: Maureen Helm <maureen.helm@nxp.com>